window is int64 instead of int

+++ sma_set_test.go
@@ -11,7 +11,7 @@ var _ = Describe("SMASet", func() {
 	var _ = (*SMA)(nil)
 
 	It("should manage a set of Simple Moving Average calculators", func() {
-		subject := NewSMASet([]int{2, 3, 4})
+		subject := NewSMASet([]int64{2, 3, 4})
 
 		// zero until values added:
 		Expect(subject.Avg()).To(Equal([]float64{0, 0, 0}))
